+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Show node details on snapshots
 - Improve performance of snapshots rendering by lazy processing the elements
 - Fix bug when sorting maps with irregular values
+- Fix bug on trace that made tooltip stay on screen
 
 ## [2021.04.22]
 - In the request tab, recent requests now show on top of the list